+++ apps/api/Program.cs
@@ -22,6 +22,7 @@ using Mohandseto.Api.Application.AdminQuotes;
 using Mohandseto.Api.Application.AdminContent;
 using Mohandseto.Api.Application.AdminInventory;
 using Mohandseto.Api.Application.AdminProcurement;
+using Mohandseto.Api.Application.AdminCrm;
 using Mohandseto.Api.Infrastructure;
 using Serilog;
 
@@ -93,6 +94,7 @@ builder.Services.AddScoped<AdminQuoteService>();
 builder.Services.AddScoped<AdminContentService>();
 builder.Services.AddScoped<AdminInventoryService>();
 builder.Services.AddScoped<AdminProcurementService>();
+builder.Services.AddScoped<AdminCrmService>();
 builder.Services.AddHostedService<ContentDispatchWorker>();
 builder.Services.AddScoped<CustomizationService>();
 builder.Services.AddSingleton<ISmsSender, ConsoleSmsSender>();